Removed weird Windows-only 'fix' that causes the URL to be broken when using Opera GX
